adk_chess/pyproject.toml
[project]
name = "adk-chess-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a>=0.44",
    "a2a-server>=0.5.8",
    "agentops>=0.4.20",
    "cairosvg>=2.8.2",
    "google-adk[a2a]>=1.12.0",
    "gradio>=5.43.1",
    "pillow>=11.3.0",
    "python-chess>=1.999",
    "python-dotenv>=1.1.1",
]
